added daily login reward and did exchange menu

+++ src/thederpgamer/edencore/utils/DateUtils.java
@@ -11,13 +11,13 @@ import java.util.Date;
  */
 public class DateUtils {
 
-    public static int getAgeDays(Date date) {
+    public static float getAgeDays(Date date) {
         Date current = new Date(System.currentTimeMillis());
         long difference = Math.abs(current.getTime() - date.getTime());
-        return (int) (difference / (1000 * 60 * 60 * 24));
+        return ((float) (difference / (1000 * 60 * 60 * 24)));
     }
 
-    public static int getAgeDays(long time) {
+    public static float getAgeDays(long time) {
         return getAgeDays(new Date(time));
     }
